change vaccine to accept injected specifier
